+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-## NEXT
+## 5.0.0-beta.3
+
+**BREAKING CHANGES:**
+
+* Flutter 3.19.0 is now required.
+* [iOS] iOS 12.0 is now the minimum supported iOS version.
+* [iOS] Adds a Privacy Manifest.
+
+Bugs fixed:
+* Fixed an issue where the camera preview and barcode scanner did not work the second time on web.
+
+Improvements:
+* [web] Migrates to extension types. (thanks @koji-1009 !)
+
+## 5.0.0-beta.2
 
 Bugs fixed:
 * Fixed an issue where the scan window was not updated when its size was changed. (thanks @navaronbracke !)
@@ -7,7 +21,6 @@ Bugs fixed:
 
 **BREAKING CHANGES:**
 
-* Flutter 3.19.0 is now required.
 * The `width` and `height` of `BarcodeCapture` have been removed, in favor of `size`.
 * The `raw` attribute is now `Object?` instead of `dynamic`, so that it participates in type promotion.
 * The `MobileScannerArguments` class has been removed from the public API, as it is an internal type.